requirements.txt
paramiko
sshtunnel
mysql-connector-python
python-dotenv